refactor: Replace DateHeader with WeekDateNavigator in TimelineView and TaskCard for improved date navigation
