+++ README.md
@@ -45,7 +45,10 @@ FocusZone/
 │   │   ├── AppPicker.swift
 │   │   ├── AppTextField.swift
 │   │   ├── AppToggle.swift
-│   │   ├── DateHeader.swift
+│   │   ├── Date/
+│   │   │   ├── WeekDateNavigator.swift
+│   │   │   ├── DateDayView.swift
+│   │   │   ├── DatePickerSheet.swift
 │   │   ├── DateSelector.swift
 │   │   ├── MainTabView.swift
 │   │   ├── TaskTimer.swift
